+++ frontend/techpick/src/utils/createSearchSelectOptions.ts
@@ -1,10 +1,10 @@
-import { SearchSelectOption } from '../types';
+import type { SearchSelectOptionType } from '@/types/SearchSelectOptionType';
 
 export const createSearchSelectOptions = (
   list: Array<{ id: number; name: string }>,
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  filterCondition: (item: any) => boolean = () => true
-): SearchSelectOption[] =>
+  // biome-ignore lint/suspicious/noExplicitAny: <explanation>
+  filterCondition: (item: any) => boolean = () => true,
+): SearchSelectOptionType[] =>
   list.filter(filterCondition).map((item) => ({
     value: item.id,
     label: item.name,
