+++ frontend/techpick/src/utils/findTagById.ts
@@ -1,4 +1,4 @@
-import type { TagType } from '@/types';
+import type { TagType } from '@/types/TagType';
 
 export const findTagById = (id: number, tagList: TagType[]) => {
   return tagList.find((tag) => tag.id === id);
