+++ frontend/techpick/src/utils/formatDateString.ts
@@ -3,5 +3,5 @@ export const formatDateString = (dateStringFromServer: string) => {
   const year = date.getFullYear(); // 2023
   const month = (date.getMonth() + 1).toString().padStart(2, '0'); // 06
   const day = date.getDate().toString().padStart(2, '0'); // 18
-  return year + '-' + month + '-' + day;
+  return `${year}-${month}-${day}`;
 };
