+++ frontend/techpick/src/utils/hasIndex.ts
