+++ frontend/techpick/src/utils/isSelectionActive.ts
